Review_bassed_journalling/requirements.txt
openai
langchain
langchain-openai
python-dotenv
faiss-cpu
pypdf
tiktoken
streamlit
